+++ parcels/_index_search.py
@@ -10,7 +10,7 @@ from parcels.statuscodes import _raise_time_extrapolation_error
 if TYPE_CHECKING:
     from parcels.xgrid import XGrid
 
-    from .field import Field
+    from parcels._core.field import Field
 
 
 GRID_SEARCH_ERROR = -3
