+++ parcels/fieldset.py
@@ -13,7 +13,7 @@ from parcels._core.utils.time import get_datetime_type_calendar
 from parcels._core.utils.time import is_compatible as datetime_is_compatible
 from parcels._logger import logger
 from parcels._typing import Mesh
-from parcels.field import Field, VectorField
+from parcels._core.field import Field, VectorField
 from parcels.converters import Geographic, GeographicPolar
 from parcels._core.xgrid import _DEFAULT_XGCM_KWARGS, XGrid
