+++ parcels/interaction/interactionkernel.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 import numpy as np
 
 from parcels._compat import MPI
-from parcels.field import VectorField
+from parcels._core.field import VectorField
 from parcels.kernel import BaseKernel
 from parcels.statuscodes import StatusCode
